Move filter logic to ViewModel
