+++ app/src/main/java/org/wikipedia/watchlist/WatchlistViewModel.kt
@@ -11,7 +11,9 @@ import org.wikipedia.WikipediaApp
 import org.wikipedia.dataclient.ServiceFactory
 import org.wikipedia.dataclient.WikiSite
 import org.wikipedia.dataclient.mwapi.MwQueryResult
+import org.wikipedia.page.Namespace
 import org.wikipedia.settings.Prefs
+import java.util.*
 
 class WatchlistViewModel : ViewModel() {
 
@@ -19,12 +21,37 @@ class WatchlistViewModel : ViewModel() {
         _uiState.value = UiState.Error(throwable)
     }
 
-    val watchlistItems = ArrayList<MwQueryResult.WatchlistItem>()
+    private val watchlistItems = mutableListOf<MwQueryResult.WatchlistItem>()
+    val finalList = mutableListOf<Any>()
     var displayLanguages = WikipediaApp.instance.languageState.appLanguageCodes.filterNot { Prefs.watchlistDisabledLanguages.contains(it) }
     var filterMode = WatchlistFragment.FILTER_MODE_ALL
 
     private val _uiState = MutableStateFlow(UiState())
     val uiState = _uiState
+
+    fun updateList() {
+        finalList.clear()
+        finalList.add("") // placeholder for header
+
+        val calendar = Calendar.getInstance()
+        var curDay = -1
+
+        for (item in watchlistItems) {
+            if (filterMode == WatchlistFragment.FILTER_MODE_ALL ||
+                (filterMode == WatchlistFragment.FILTER_MODE_PAGES && Namespace.of(item.ns).main()) ||
+                (filterMode == WatchlistFragment.FILTER_MODE_TALK && Namespace.of(item.ns).talk()) ||
+                (filterMode == WatchlistFragment.FILTER_MODE_OTHER && !Namespace.of(item.ns).main() && !Namespace.of(item.ns).talk())) {
+
+                calendar.time = item.date
+                if (calendar.get(Calendar.DAY_OF_YEAR) != curDay) {
+                    curDay = calendar.get(Calendar.DAY_OF_YEAR)
+                    finalList.add(item.date)
+                }
+
+                finalList.add(item)
+            }
+        }
+    }
 
     fun fetchWatchlist() {
         viewModelScope.launch(handler) {
